Refactor osrs_ge_api
